% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utils.R
\name{decimal_date_local}
\alias{decimal_date_local}
\title{Convert dates to decimal years (day-of-year convention)}
\usage{
decimal_date_local(x)
}
\arguments{
\item{x}{\code{Date} vector (or coercible).}
}
\value{
Numeric vector of decimal years.
}
\description{
Converts dates to decimal years as \code{year + (day_of_year - 1)/365},
the convention used throughout the WAI vintage handling. Unlike
\code{lubridate::decimal_date()}, this does not account for leap years.
}
\examples{
decimal_date_local(as.Date("2020-03-07"))

}
