% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_world}
\alias{hello_world}
\title{Say hello}
\usage{
hello_world(name = "world")
}
\arguments{
\item{name}{A character string with the name to greet. Defaults to
\code{"world"}.}
}
\value{
A character string containing the greeting.
}
\description{
A simple example function that returns a friendly greeting. Use it as a
template for writing your own package functions.
}
\examples{
hello_world()
hello_world("Philipp")
}
