% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utils.R
\name{is_crisis_period}
\alias{is_crisis_period}
\title{Flag dates falling into the crisis periods}
\usage{
is_crisis_period(date_vec)
}
\arguments{
\item{date_vec}{\code{Date} vector (or coercible).}
}
\value{
Logical vector.
}
\description{
Marks dates inside the financial crisis (2008-07-07 to 2009-09-28) or
the Covid-19 crisis (2020-01-01 to 2021-12-28).
}
\examples{
is_crisis_period(as.Date(c("2015-01-01", "2020-06-01")))

}
