% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-config.R
\name{latest_fit_file}
\alias{latest_fit_file}
\title{Find the newest fit file up to a cutoff date}
\usage{
latest_fit_file(folder, cutoff_decimal)
}
\arguments{
\item{folder}{Directory containing \verb{fit_<decimal-date>.Rda} files.}

\item{cutoff_decimal}{Numeric decimal date; only fits at or before this
cutoff are considered.}
}
\value{
Full path of the selected fit file.
}
\description{
Find the newest fit file up to a cutoff date
}
