% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcast.R
\name{retrieve_nowcast}
\alias{retrieve_nowcast}
\title{Extract the nowcast from a fit object}
\usage{
retrieve_nowcast(fit, model)
}
\arguments{
\item{fit}{A fit object from \code{\link[=run_ar]{run_ar()}} or \code{\link[=run_wai_adj]{run_wai_adj()}}.}

\item{model}{Character, \code{"ar"} or \code{"wai"}.}
}
\value{
The nowcast value.
}
\description{
Extract the nowcast from a fit object
}
